--- homework/homework-1.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Cursos da Alura</title>
-</head>
-<body>
-    <ul>
-        <li>Kotlin</li>
-        <li>Docker</li>
-        <li>Java</li>
-        <li>Spring</li>
-    </ul>
-</body>
-</html>
