+++ modulo1/homework01/src/Contato.java
@@ -8,7 +8,6 @@ public class Contato {
         this.telefone = telefone;
         this.tipoContato = tipoContato;
     }
-
     public void imprimirContato() {
         System.out.println("Descrição : " + descricao + "\nTelefone : " + telefone + "\nTipo de Contatp : " + tipoContato);
     }
